# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Rip",
        "title": "Rip",
        "contact": {
            "name": "Rip API Support",
            "email": "jomolun@yandex.com"
        },
        "version": "1.0.1"
    },
    "host": "localhost:80",
    "basePath": "/api/",
    "paths": {
...
        "/api/encryptDecryptRequest/filter": {
            "get": {
                "description": "Get requests filtered by status, start and end date",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EncryptDecryptRequest"
                ],
                "summary": "Filter EncryptDecryptRequests",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start Date",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End Date",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data Service Filtered",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/rip_internal_pkg_repo.EncryptDecryptRequestView"
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/encryptDecryptRequest/form/{id}": {
            "put": {
                "description": "Form EncryptDecryptRequest by moderator",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EncryptDecryptRequest"
                ],
                "summary": "Form EncryptDecryptRequest by moderator",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "EncryptDecryptRequest ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "EncryptDecryptRequest and DataServices",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/encryptDecryptRequest/update_moderator/{id}": {
            "put": {
                "description": "Update EncryptDecryptRequest by moderator",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EncryptDecryptRequest"
                ],
                "summary": "Update EncryptDecryptRequest by moderator",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "EncryptDecryptRequest ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Action: finish or reject",
                        "name": "action",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Finished/Rejected",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/encryptDecryptRequest/{req_id}": {
            "get": {
                "description": "Get EncryptDecryptRequest by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EncryptDecryptRequest"
                ],
                "summary": "Get EncryptDecryptRequest by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "EncryptDecryptRequest ID",
                        "name": "req_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Draft EncryptDecryptRequest and Data Services",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Delete EncryptDecryptRequest by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EncryptDecryptRequest"
                ],
                "summary": "Delete EncryptDecryptRequest by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "EncryptDecryptRequest ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/encryptDecryptRequest/{req_id}/delete/{data_id}": {
            "delete": {
                "description": "Delete DataService from EncryptDecryptRequest",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EncryptDecryptRequest"
                ],
                "summary": "Delete DataService from EncryptDecryptRequest",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "EncryptDecryptRequest ID",
                        "name": "req_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "DataService ID",
                        "name": "data_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
...
        "rip_internal_pkg_repo.EncryptDecryptRequestView": {
            "type": "object",
            "properties": {
                "creationDate": {
                    "type": "string"
                },
                "creator": {
                    "type": "string"
                },
                "finishDate": {
                    "type": "string"
                },
                "formDate": {
                    "type": "string"
                },
                "moderator": {
                    "type": "string"
                },
                "requestID": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/rip_internal_pkg_repo.Status"
                }
            }
        },
        "rip_internal_pkg_repo.Status": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5
            ],
            "x-enum-varnames": [
                "Draft",
                "Deleted",
                "Formed",
                "Finished",
                "Rejected",
                "UnknownStatus"
            ]
        },
...
    }
}